Added 404 for get, update and delete

+++ src/main/java/br/com/amsj/forum/controller/TopicoController.java
@@ -2,6 +2,7 @@ package br.com.amsj.forum.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -46,11 +47,15 @@ public class TopicoController {
 	}
 	
 	@GetMapping("/{id}")
-	public TopicoDetailDto detail(@PathVariable Long id){
+	public ResponseEntity<TopicoDetailDto> detail(@PathVariable Long id){
 		
-		Topico topico = topicoRepository.getOne(id);
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			return ResponseEntity.ok(new TopicoDetailDto(optional.get()));	
+		}
 		
-		return new TopicoDetailDto(topico);
+		return ResponseEntity.notFound().build();
+		
 	}
 	
 	@GetMapping("/byCursoName/{nomeCurso}")
@@ -77,18 +82,24 @@ public class TopicoController {
 	@Transactional
 	public ResponseEntity<TopicoDto> update(@PathVariable Long id, @RequestBody @Valid TopicoUpdateForm topicoUpdateForm) {
 		
-		Topico topico = topicoUpdateForm.update(id, topicoRepository);
-		
-		return ResponseEntity.ok(new TopicoDto(topico));
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			Topico topico = topicoUpdateForm.update(id, topicoRepository);		
+			return ResponseEntity.ok(new TopicoDto(topico));
+		}
+		return ResponseEntity.notFound().build();
 	}
 	
 	@DeleteMapping("/{id}")
 	@Transactional
 	public ResponseEntity<?> delete(@PathVariable Long id) {
 		
-		topicoRepository.deleteById(id);
-		
-		return ResponseEntity.ok().build();
+		Optional<Topico> optional = topicoRepository.findById(id);
+		if(optional.isPresent()) {
+			topicoRepository.deleteById(id);
+			return ResponseEntity.ok().build();
+		}
+		return ResponseEntity.notFound().build();
 	}
 
 }
